requirements-dev.txt
curl_cffi
commonX
PyYAML
Pillow
psutil
pycryptodome
requests
plugin_jm_server
zhconv
Pillow